services/ingestion/requirements.txt
boto3==1.34.135
playwright==1.47.0
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dateutil==2.9.0.post0
requests==2.32.3
tenacity==8.5.0
vaderSentiment==3.3.2
langdetect==1.0.9
beautifulsoup4==4.12.3
feedparser==6.0.11
